Freeze protection explicitly forbidden in non-windows environments

+++ mpqprt/mpq/comp/huffman.h
@@ -15,12 +15,14 @@
 //-----------------------------------------------------------------------------
 // Defines
 
+#include <stddef.h>
+
 #define INSERT_ITEM    1                    
 #define SWITCH_ITEMS   2                    // Switch the item1 and item2
 
 #define PTR_NOT(ptr)  (THTreeItem *)(~(unsigned long)(ptr))
 #define PTR_PTR(ptr)  ((THTreeItem *)(ptr))
-#define PTR_INT(ptr)  (long)(ptr)
+#define PTR_INT(ptr)  (ptrdiff_t)(ptr)
 
 #ifndef NULL
 #define NULL 0
